Isolate rendering allocation tests
Move the allocation-specific expectations and render cases into a dedicated test file.


Copilot-Session: b4cd86e1-c7a5-453e-bb5a-8cc2a9d6edd0
